+++ backtesting/_util.py
@@ -1,6 +1,6 @@
 import warnings
-from typing import Dict, List, Optional, Sequence, Union
 from numbers import Number
+from typing import Dict, List, Optional, Sequence, Union, cast
 
 import numpy as np
 import pandas as pd
@@ -56,6 +56,16 @@ class _Array(np.ndarray):
             self.name = getattr(obj, 'name', '')
             self._opts = getattr(obj, '_opts', {})
 
+    # Make sure properties name and _opts are carried over
+    # when (un-)pickling.
+    def __reduce__(self):
+        value = super().__reduce__()
+        return value[:2] + (value[2] + (self.__dict__,),)
+
+    def __setstate__(self, state):
+        self.__dict__.update(state[-1])
+        super().__setstate__(state[:-1])
+
     def __bool__(self):
         try:
             return bool(self[-1])
@@ -75,13 +85,14 @@ class _Array(np.ndarray):
     @property
     def s(self) -> pd.Series:
         values = np.atleast_2d(self)
-        return pd.Series(values[0], index=self._opts['data'].index, name=self.name)
+        index = self._opts['index'][:values.shape[1]]
+        return pd.Series(values[0], index=index, name=self.name)
 
     @property
     def df(self) -> pd.DataFrame:
         values = np.atleast_2d(np.asarray(self))
-        df = pd.DataFrame(values.T, index=self._opts['data'].index,
-                          columns=[self.name] * len(values))
+        index = self._opts['index'][:values.shape[1]]
+        df = pd.DataFrame(values.T, index=index, columns=[self.name] * len(values))
         return df
 
 
@@ -118,10 +129,11 @@ class _Data:
         self.__cache.clear()
 
     def _update(self):
-        self.__arrays = {col: _Array(arr, data=self)
+        index = self.__df.index.copy()
+        self.__arrays = {col: _Array(arr, index=index)
                          for col, arr in self.__df.items()}
         # Leave index as Series because pd.Timestamp nicer API to work with
-        self.__arrays['__index'] = self.__df.index.copy()
+        self.__arrays['__index'] = index
 
     def __repr__(self):
         i = min(self.__i, len(self.__df) - 1)
@@ -141,14 +153,14 @@ class _Data:
     @property
     def pip(self) -> float:
         if self.__pip is None:
-            self.__pip = 10**-np.median([len(s.partition('.')[-1])
-                                         for s in self.__arrays['Close'].astype(str)])
+            self.__pip = float(10**-np.median([len(s.partition('.')[-1])
+                                               for s in self.__arrays['Close'].astype(str)]))
         return self.__pip
 
     def __get_array(self, key) -> _Array:
         arr = self.__cache.get(key)
         if arr is None:
-            arr = self.__cache[key] = self.__arrays[key][:self.__i]
+            arr = self.__cache[key] = cast(_Array, self.__arrays[key][:self.__i])
         return arr
 
     @property